Move the new Security helper to the root of SecurityBundle
